+++ src/test/php/de/thekid/cas/unittest/TestingUsers.php
@@ -8,8 +8,8 @@ class TestingUsers extends Users {
   private $backing= [];
 
   public function __construct(array $users= []) {
-    foreach ($users as $username => $password) {
-      $this->backing[$username]= $this->hash($password);
+    foreach ($users as $username => $details) {
+      $this->backing[$username]= ['hash' => $this->hash($details['password']), 'tokens' => $details['tokens'] ?? []];
     }
   }
 
@@ -20,13 +20,13 @@ class TestingUsers extends Users {
 
   /** Returns a user by a given username */
   public function named(string $username): ?User {
-    $hash= $this->backing[$username] ?? null;
-    return $hash ? new User($username, $hash, []) : null;
+    $user= $this->backing[$username] ?? null;
+    return $user ? new User($username, $user['hash'], $user['tokens']) : null;
   }
 
   /** Creates a new user with a given username and password. */
   public function create(string $username, string|Secret $password): User {
-    $this->backing[$username]= $this->hash($password);
+    $this->backing[$username]= ['hash' => $this->hash($password), 'tokens' => []];
     return new User($username, $this->backing[$username], []);
   }
 
@@ -39,6 +39,16 @@ class TestingUsers extends Users {
   /** Changes a user's password. */
   public function password(string|User $user, string|Secret $password): void {
     $username= $user instanceof User ? $user->username() : $user;
-    $this->backing[$username]= $this->hash($password);
+    $this->backing[$username]['hash']= $this->hash($password);
+  }
+
+  public function newToken(string|User $user, string $name, string|Secret $secret) {
+    $username= $user instanceof User ? $user->username() : $user;
+    $this->backing[$username]['tokens'][$name]= $secret instanceof Secret ? $secret->reveal() : $secret;
+  }
+
+  public function removeToken(string|User $user, string $name) {
+    $username= $user instanceof User ? $user->username() : $user;
+    unset($this->backing[$username]['tokens'][$name]);
   }
 }